requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
langchain-chroma>=0.2.0
pypdf>=5.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.39.0
mcp>=1.2.0
